+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0
 
 * Added MultiCAS methods for update ops.
+* Added store with fetch and update.
 
 ## 0.1.0